SmartCart/backend/requirements.txt
# FastAPI and GraphQL
fastapi==0.104.1
uvicorn==0.24.0
strawberry-graphql==0.211.1
pydantic==2.5.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# AI/ML
tensorflow==2.14.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic-settings==2.1.0

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.35.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# New dependencies
asyncio==3.4.3
joblib==1.3.2
celery==5.3.6
flower==2.0.1
grafana-api-client==2.1.0 